+++ flaskr/app.py
@@ -1,3 +1,4 @@
+from vistas.vistas import VistaUsuario
 from flaskr import create_app
 from flask_restful import Api
 from .modelos import db
@@ -22,5 +23,6 @@ api.add_resource(VistaLogIn, '/logIn')
 api.add_resource(VistaAlbumsUsuario, '/usuario/<int:id_usuario>/albumes')
 api.add_resource(VistaAlbum, '/album/<int:id_album>')
 api.add_resource(VistaCancionesAlbum, '/album/<int:id_album>/canciones')
+api.add_resource(VistaUsuario, '/usuario/<int:id_usuario>')
 
 jwt = JWTManager(app)
