+++ front-end/tutorial-canciones/src/app/app.component.html
@@ -1,5 +1,5 @@
 
-<!--<app-banner></app-banner>-->
+<app-banner></app-banner>
 
 <router-outlet></router-outlet>
