--- front-end/tutorial-canciones/src/app/usuario-header/usuario-header.component.html
@@ -1,7 +1,0 @@
-
-<div class="user-image">
-    <img [src]="(usuario?.photoUrl)? usuario?.photoUrl : 'http://s3.amazonaws.com/37assets/svn/765-default-avatar.png'">
-</div>
-<div class="user-info">
-    <h3>{{usuario?.email}}</h3>
-</div>
