+++ front-end/tutorial-canciones/src/app/usuario/usuario-header/usuario-header.component.css
@@ -1,0 +1,4 @@
+div.user-image > img{
+    width: 60px;
+    border-radius: 60px;
+    }
