+++ front-end/tutorial-canciones/src/app/usuario/usuario-header/usuario-header.component.html
@@ -1,0 +1,7 @@
+
+<div class="user-image">
+    <img [src]="'http://s3.amazonaws.com/37assets/svn/765-default-avatar.png'">
+</div>
+<div class="user-info">
+    <h6>{{usuario?.nombre}}</h6>
+</div>
