+++ front-end/tutorial-canciones/src/app/usuario/usuario-header/usuario-header.component.ts
@@ -1,0 +1,26 @@
+import { Component, OnInit, Input} from '@angular/core';
+import { UsuarioService } from '../usuario.service';
+import { Usuario } from '../usuario';
+
+@Component({
+  selector: 'app-usuario-header',
+  templateUrl: './usuario-header.component.html',
+  styleUrls: ['./usuario-header.component.css']
+})
+export class UsuarioHeaderComponent implements OnInit {
+ 
+  @Input() usuario_id: number;
+  usuario : Usuario
+
+  constructor(private usuarioService: UsuarioService) { }
+
+  ngOnInit(): void {
+    this.usuarioService.userGet(this.usuario_id)
+    .subscribe(
+      usuario => {
+        this.usuario = usuario;
+      }
+    );
+  }
+
+}
