+++ front-end/tutorial-canciones/src/app/usuario/usuario.service.ts
@@ -1,6 +1,7 @@
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { Observable } from 'rxjs';
+import { Usuario } from './usuario';
 
 @Injectable({
     providedIn: 'root'
@@ -17,5 +18,9 @@ export class UsuarioService {
 
     userSignUp(nombre: string, contrasena: string): Observable<any>{
         return this.http.post<any>(`${this.backUrl}/signin`, {"nombre": nombre, "contrasena": contrasena})
+    }
+
+    userGet(usuario_id: number): Observable<Usuario>{
+        return this.http.get<Usuario>(`${this.backUrl}/usuario/${usuario_id}`)
     }
 }
