+++ front-end/tutorial-canciones/src/styles.css
@@ -1,5 +1,1 @@
 /* You can add global styles to this file, and also import other style files */
-div.user-image > img{
-    width: 80px;
-    border-radius: 50px;
-    }
